Refactored, added method, and broke tests into appropriate files
